A Python program running the KFLOP Dynomotion
This project contains 3 programs; 1) a sample program running the KFLOP Dynomotion Motion Controller.  2) Web Services program to Teleoperate the KFLOP motion controller from a Web Browser.  3) HTML front-end for the Web Services

These programs are as-is.  Do not use them in mission critical applications such as but not limited to life support, avionics, or space exploration